// Web/react-app/src/models/ITeapotRequest.ts
interface ITeapotRequest {
    name: string;
    quantity: number;
    color: string;
    bodyMaterial: string;
    power: number;
    price: number;
    imgName: string;
    volume: number;
    warrantyInMonths: number;
    functions: string;
    weight: number;
    companyId: number;
    manufacturerCountry: string;
}

export default ITeapotRequest;